merge pjsip into the main package
Other core module, res_rtp_asterisk.so, requires res_pjproject.so, so
asterisk-pjsip would have to be installed even when the old chan_sip is
used.
We don't build with system pjproject any more, so the 'pjsip' subpackage
doesn't pull any dependencies now, so including it in the main package
won't hurt.